+++ gazebo/physics/UserCmdManager.cc
@@ -210,6 +210,7 @@ void UserCmdManager::OnUserCmdMsg(ConstUserCmdPtr &_msg)
 
       auto wrenchPub = this->dataPtr->node->Advertise<msgs::Wrench>(topicName);
       wrenchPub->Publish(_msg->wrench());
+      wrenchPub->Fini();
 
       break;
     }
@@ -358,5 +359,3 @@ void UserCmdManager::PublishCurrentStats()
 
   this->dataPtr->userCmdStatsPub->Publish(statsMsg);
 }
-
-